Moved all logic into a processor class (refactoring step 1).
